Add python package
Named "us_nicknames" because I wanted to denote that this is
only US names, and it is pretty short.

Uses hatch as the build tool.
Generally pretty straightforwar

This does remove python-parser.py, so it will break anyone who uses this
from source. But if the API is pretty much the same, so the transition
should be easy.

Now the Denormalizer is easier to use because you can supply your own
lookup directly instead of needing to load it from a file.
The arg name to __init__ changed from `filename`
to `lookup` because that's more semantic, it could break someone.