Revert names.csv to pre-normalization
We ran the normalize.py in
https://github.com/carltonnorthern/nickname-and-diminutive-names-lookup/pull/37
and commited the result, but that actually didn't
follow the schema of the file that we decided
upon.

So go back to the original schema, so we can
normalize again with the updated script that accounts for
the first name in the line being the canonical
one.