Overhaul python parser to respect assymetry
Now we care about the canonical <-> nickname
relationship discussed in
https://github.com/carltonnorthern/nickname-and-diminutive-names-lookup/pull/37

I didn't add the `get_aliases` method because
I wanted to keep the API small, and users could do that if they wanted.

Required keyword args to the __init__ so you couldn't accidentally
try to create from a csv file path. For ease of use though you
can not provide anything, and it still loads the data
from the CSV as before.

This is slightly breaking again, but again should be easy to adjust.