Overhaul generate_sql.py, allow parameterization
Now
- you can choose a file path for output
- reading the CSV is more intuitive
- template strings are more intuitive

Some parts of this are more intuitive, but this also is a bit more
complicated. I was finding it hard to
stay DRY though, because I needed
to split up the generation
into two different functions,
one for normalized and one for non-normalized. Perhaps could make this
less complicated.